Edit navlinks
