+++ src/widgets/navBar/components/menuItem.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 
 import { IMenuItem } from '../api/menuItemsTypes';
-import MenuLink from './menuLink';
+import MenuLink from './MenuLink/menuLink';
 
 interface MenuItemProps {
     item: IMenuItem;
